api-fast/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
blinker==1.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
defusedxml==0.8.0rc2
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
fastapi==0.110.2
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.0
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
mysql-connector-python-rf==2.2.2
nltk==3.8.1
oauthlib==3.2.2
passlib==1.7.4
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
python-jose==3.3.0
python3-openid==3.2.0
regex==2024.4.28
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
social-auth-core==4.5.4
starlette==0.37.2
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
Werkzeug==3.0.2
